+++ src/stores/tokenInfoStore.ts
@@ -6,8 +6,13 @@ import type { GetWebTokenInfo, TokenBalanceRaw } from '@/mew_api/types'
 import type { ComputedRef } from 'vue'
 
 export const useTokenInfoStore = defineStore('tokenInfoStore', () => {
-  const tokenStore = useLocalStorage<(TokenBalanceRaw | GetWebTokenInfo)[]>('tokenInfo', [], { mergeDefaults: true })
-  const tokenInfo: ComputedRef<TokenBalanceRaw | GetWebTokenInfo | null> = computed(() => tokenStore.value[0] || null)
+  const tokenStore = useLocalStorage<(TokenBalanceRaw | GetWebTokenInfo)[]>(
+    'tokenInfo',
+    [],
+    { mergeDefaults: true },
+  )
+  const tokenInfo: ComputedRef<TokenBalanceRaw | GetWebTokenInfo | null> =
+    computed(() => tokenStore.value[0] || null)
 
   const setTokenInfo = (value: TokenBalanceRaw | GetWebTokenInfo) => {
     tokenStore.value = [value]
@@ -20,6 +25,6 @@ export const useTokenInfoStore = defineStore('tokenInfoStore', () => {
   return {
     tokenInfo,
     setTokenInfo,
-    removeTokenInfo
+    removeTokenInfo,
   }
 })