+++ CHANGELOG.md
@@ -1,12 +1,16 @@
+#### Feat
+- Add filter low fiat balance tokens [#5333](https://github.com/MyEtherWallet/MyEtherWallet/pull/5333)
+
+
 ### v7.0.2
 
-# Feat
+#### Feat
 
 - Swap and Trade Tokenized stock rewards [#5327](https://github.com/MyEtherWallet/MyEtherWallet/pull/5327)
 
 ### v7.0.1
 
-# Feat
+#### Feat
 
 - added intercom support [#5324](https://github.com/MyEtherWallet/MyEtherWallet/pull/5324)