+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Add blinking info for approval on ledger or trezor [#5331](https://github.com/MyEtherWallet/MyEtherWallet/pull/5331)
 - store selected pair in swap and trade, and set it once user navigates back [#5335](https://github.com/MyEtherWallet/MyEtherWallet/pull/5335)
+- fix error when watchlist tokens are cleared [#5336](https://github.com/MyEtherWallet/MyEtherWallet/pull/5336)
 
 ### v7.0.2