+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Fix
 
 - fix metamask error on cancelling send and fix bug when disconnecting through metamask [#5337](https://github.com/MyEtherWallet/MyEtherWallet/pull/5337)
-- handle shit token send error [#5330](https://github.com/MyEtherWallet/MyEtherWallet/pull/5330)
+- handle crap token send error [#5330](https://github.com/MyEtherWallet/MyEtherWallet/pull/5330)
 
 # Feat