+++ CHANGELOG.md
@@ -1,9 +1,23 @@
+### v7.0.3
+
+### Fix
+
+- fix metamask error on cancelling send and fix bug when disconnecting through metamask [#5337](https://github.com/MyEtherWallet/MyEtherWallet/pull/5337)
+- handle shit token send error [#5330](https://github.com/MyEtherWallet/MyEtherWallet/pull/5330)
+
+# Feat
+
+- Add blinking info for approval on ledger or trezor [#5331](https://github.com/MyEtherWallet/MyEtherWallet/pull/5331)
+- store selected pair in swap and trade, and set it once user navigates back [#5335](https://github.com/MyEtherWallet/MyEtherWallet/pull/5335)
+- fix error when watchlist tokens are cleared [#5336](https://github.com/MyEtherWallet/MyEtherWallet/pull/5336)
+
 ### v7.0.2
 
 # Feat
 
 - Swap and Trade Tokenized stock rewards [#5327](https://github.com/MyEtherWallet/MyEtherWallet/pull/5327)
 - Balance Refresh and Tx has for rewards [#5334](https://github.com/MyEtherWallet/MyEtherWallet/pull/5334)
+- Account eligibility [#5339](https://github.com/MyEtherWallet/MyEtherWallet/pull/5339)
 
 ### v7.0.1